HS-583: Add TaskSet for Snmp Detector
