+++ shared-lib/task-set-service/pom.xml
@@ -18,7 +18,6 @@
 
     <modules>
         <module>api</module>
-        <module>client</module>
         <module>ignite-client</module>
     </modules>
